keraban/wsgi.py
"""
WSGI config for keraban project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/2.0/howto/deployment/wsgi/
"""

import os
import dotenv

from django.core.wsgi import get_wsgi_application

# load environment settings from file if there is a file `.env` at root of project
# https://github.com/jpadilla/django-dotenv
dotenv.read_dotenv(os.path.join(os.path.dirname(os.path.dirname(__file__)), '.env'))

os.environ.setdefault("DJANGO_SETTINGS_MODULE", "keraban.settings.dev")

application = get_wsgi_application()
