requirements.txt
beautifulsoup4==4.6.0
brotlipy==0.7.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Django==2.0.5
django-dotenv==1.4.2
django-modelcluster==4.1
django-taggit==0.22.2
django-treebeard==4.3
djangorestframework==3.8.2
draftjs-exporter==2.0.0
html5lib==0.999999999
idna==2.6
libsass==0.14.5
Pillow==5.1.0
pycparser==2.18
pytz==2018.4
requests==2.18.4
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
wagtail==2.0.1
wagtailfontawesome==1.1.3
webencodings==0.5.1
whitenoise==3.3.1
Willow==1.1
